Big modification for project
Modified structure of the project. Added custom controller and service
appropriately for it (now it is just for test, in the future it must be
for  right panel of liked reddits). Added constants in the project.
Added one directive for image (now it is not uses anywhere).  Simple
modification for the prj styles - it's better now.